Extracted floor into a component
